+++ pypeit/core/wavecal/autoid.py
@@ -19,7 +19,6 @@ from pypeit.core.wavecal import waveio
 from pypeit.core.wavecal import patterns
 from pypeit.core.wavecal import fitting
 from pypeit.core.wavecal import wvutils
-from pypeit.core import qa
 from pypeit.core import arc
 
 from pypeit.core import pca
@@ -28,6 +27,214 @@ from pypeit import utils
 from pypeit import msgs
 from pypeit import debugger
 from matplotlib import pyplot as plt
+from matplotlib import gridspec
+from matplotlib.backends.backend_pdf import PdfPages
+from matplotlib.patches import Patch
+
+
+def arc_fit_qa(fit, outfile=None, ids_only=False, title=None):
+    """
+    QA for Arc spectrum
+
+    Parameters
+    ----------
+    setup: str
+      For outfile
+    fit : dict
+    arc_spec : ndarray
+      Arc spectrum
+    outfile : str, optional
+      Name of output file
+      or 'show' to show on screen
+    """
+
+    plt.rcdefaults()
+    plt.rcParams['font.family']= 'times new roman'
+
+    arc_spec = fit['spec']
+
+    # Begin
+    plt.close('all')
+    if ids_only:
+        nrows, ncols = 1,1
+        figsize =(11,8.5)
+        idfont = 'small'
+    else:
+        nrows, ncols = 2,2
+        if outfile is None:
+            figsize = (16,8)
+            idfont = 'small'
+        else:
+            figsize = (8,4)
+            idfont = 'xx-small'
+    fig = plt.figure(figsize=figsize)
+    gs = gridspec.GridSpec(nrows,ncols)#, figure = fig)
+
+
+    # Simple spectrum plot
+    ax_spec = plt.subplot(gs[:,0])
+    ax_spec.plot(np.arange(len(arc_spec)), arc_spec)
+    ymin, ymax = np.min(arc_spec), np.max(arc_spec)
+    ysep = ymax*0.03
+    mask = fit['mask']
+    pixel_fit = fit['pixel_fit'][mask]
+    wave_fit = fit['wave_fit'][mask]
+    ions = fit['ions'][mask]
+    xnorm = fit['xnorm']
+    for kk, x in enumerate(pixel_fit):
+        ind_left = np.fmax(int(x)-2, 0)
+        ind_righ = np.fmin(int(x)+2,arc_spec.size-1)
+        yline = np.max(arc_spec[ind_left:ind_righ])
+        # Tick mark
+        ax_spec.plot([x,x], [yline+ysep*0.25, yline+ysep], 'g-')
+        # label
+        ax_spec.text(x, yline+ysep*1.3,'{:s} {:g}'.format(ions[kk], wave_fit[kk]), ha='center', va='bottom',size=idfont,
+                     rotation=90., color='green')
+    ax_spec.set_xlim(0., len(arc_spec))
+    ax_spec.set_ylim(1.05*ymin, ymax*1.2)
+    ax_spec.set_xlabel('Pixel')
+    ax_spec.set_ylabel('Flux')
+    if title is not None:
+        ax_spec.text(0.04, 0.93, title, transform=ax_spec.transAxes,
+                     size='x-large', ha='left')#, bbox={'facecolor':'white'})
+    if ids_only:
+        plt.tight_layout(pad=0.2, h_pad=0.0, w_pad=0.0)
+        if outfile is None:
+            plt.show()
+        else:
+            plt.savefig(outfile, dpi=800)
+        plt.close()
+        return
+
+    # Arc Fit
+    ax_fit = plt.subplot(gs[0, 1])
+    # Points
+    ax_fit.scatter(pixel_fit,wave_fit, marker='x')
+    if len(fit['xrej']) > 0:
+        ax_fit.scatter(fit['xrej'], fit['yrej'], marker='o',
+            edgecolor='gray', facecolor='none')
+    # Solution
+    xval = np.arange(len(arc_spec))
+    wave_soln = fit['wave_soln'] #utils.func_val(fit['fitc'], xval, 'legendre',minx=fit['fmin'], maxx=fit['fmax'])
+    ax_fit.plot(xval, wave_soln, 'r-')
+    xmin, xmax = 0., len(arc_spec)
+    ax_fit.set_xlim(xmin, xmax)
+    ymin,ymax = np.min(wave_soln)*.95,  np.max(wave_soln)*1.05
+    ax_fit.set_ylim((ymin, ymax))
+    ax_fit.set_ylabel('Wavelength')
+    ax_fit.get_xaxis().set_ticks([]) # Suppress labeling
+    # Stats
+    wave_soln_fit = utils.func_val(fit['fitc'], pixel_fit/xnorm, 'legendre',minx=fit['fmin'], maxx=fit['fmax'])
+    rms = np.sqrt(np.sum((wave_fit-wave_soln_fit)**2)/len(pixel_fit)) # Ang
+    dwv_pix = np.median(np.abs(wave_soln-np.roll(wave_soln,1)))
+    ax_fit.text(0.1*len(arc_spec), 0.90*ymin+(ymax-ymin),r'$\Delta\lambda$={:.3f}$\AA$ (per pix)'.format(dwv_pix), size='small')
+    ax_fit.text(0.1*len(arc_spec), 0.80*ymin+(ymax-ymin),'RMS={:.3f} (pixels)'.format(rms/dwv_pix), size='small')
+    # Arc Residuals
+    ax_res = plt.subplot(gs[1,1])
+    res = wave_fit-wave_soln_fit
+    ax_res.scatter(pixel_fit, res/dwv_pix, marker='x')
+    ax_res.plot([xmin,xmax], [0.,0], 'k--')
+    ax_res.set_xlim(xmin, xmax)
+    ax_res.set_xlabel('Pixel')
+    ax_res.set_ylabel('Residuals (Pix)')
+
+    # Finish
+    plt.tight_layout(pad=0.2, h_pad=0.0, w_pad=0.0)
+    if outfile is None:
+        plt.show()
+    else:
+        plt.savefig(outfile, dpi=400)
+    plt.close('all')
+
+    plt.rcdefaults()
+
+
+    return
+
+
+def match_qa(arc_spec, tcent, line_list, IDs, scores, outfile = None, title=None, path=None):
+    """
+    Parameters
+    ----------
+    arc_spec
+    tcent
+    line_list
+    IDs
+    scores
+    outfile
+    title
+    path
+
+    Returns
+    -------
+
+    """
+
+
+    # Plot
+    plt.figure(figsize=(11, 8.5))
+    plt.clf()
+    gs = gridspec.GridSpec(1, 1)
+    idfont = 'small'
+
+    # Simple spectrum plot
+    ax_spec = plt.subplot(gs[0])
+    ax_spec.plot(np.arange(len(arc_spec)), arc_spec, 'k')
+    ymin, ymax = np.min(arc_spec), np.max(arc_spec)
+    ysep = ymax*0.03
+    mn_yline = 1e9
+
+    # Standard IDs
+    clrs = dict(Perfect='green', Good='orange', Ok='red')
+    clrs['Very Good'] = 'blue'
+    for kk, score in enumerate(scores):
+        x = tcent[kk]
+        # Color
+        try:
+            clr = clrs[score]
+        except KeyError:
+            clr = 'gray'
+        yline = np.max(arc_spec[int(x)-2:int(x)+2])
+        mn_yline = min(mn_yline, yline)
+        # Tick mark
+        ax_spec.plot([x,x], [yline+ysep*0.25, yline+ysep], '-', color=clr)
+        if score in ['Good', 'Ok', 'Perfect', 'Very Good']:
+            # Label
+            imin = np.argmin(np.abs(line_list['wave']-IDs[kk]))
+            row = line_list[imin]
+            lbl = '{:s} {:.4f}'.format(row['ion'], row['wave'])
+            # label
+            ax_spec.text(x, yline+ysep*1.3, '{:s}'.format(lbl), ha='center', va='bottom',
+                size=idfont, rotation=90., color=clr)
+    # Overplot the line classification legened
+    clrs['Not reidentified'] ='gray'
+    legend_elements = []
+    for key, clr in clrs.items():
+        legend_elements.append(Patch(facecolor=clr, edgecolor=clr,label=key))
+
+    # Axes
+    ax_spec.set_xlim(0., len(arc_spec))
+    ax_spec.set_ylim(ymin*1.05, ymax*1.3)
+    ax_spec.set_xlabel('Pixel')
+    ax_spec.minorticks_on()
+    ax_spec.set_ylabel('Counts')
+    plt.legend(handles=legend_elements)
+    if title is not None:
+        ax_spec.text(0.04, 0.93, title, transform=ax_spec.transAxes,
+                     size='x-large', ha='left')#, bbox={'facecolor':'white'})
+    # Finish
+    plt.tight_layout(pad=0.2, h_pad=0.0, w_pad=0.0)
+    if outfile is None:
+        plt.show()
+    else:
+        pp = PdfPages(outfile)
+        pp.savefig(bbox_inches='tight')
+        pp.close()
+
+    plt.close()
+    return
+
+
 
 
 def basic(spec, lines, wv_cen, disp, sigdetect=20.,nonlinear_counts = 1e10,
@@ -261,7 +468,7 @@ def semi_brute(spec, lines, wv_cen, disp, sigdetect=30., nonlinear_counts = 1e10
     # Plot
     if outroot is not None:
         tmp_list = table.vstack([line_lists,unknwns])
-        qa.match_qa(spec, cut_tcent, tmp_list, best_dict['IDs'], best_dict['scores'], outfile = outroot+'.pdf')
+        match_qa(spec, cut_tcent, tmp_list, best_dict['IDs'], best_dict['scores'], outfile = outroot+'.pdf')
         msgs.info("Wrote: {:s}".format(outroot+'.pdf'))
 
     # Fit
@@ -591,7 +798,7 @@ def reidentify(spec, spec_arxiv_in, wave_soln_arxiv_in, line_list, nreid_min, de
         plt.legend()
         plt.show()
         # QA Plot ofthe reidentifications
-        qa.match_qa(spec_cont_sub, detections, line_list, patt_dict_slit['IDs'], patt_dict_slit['scores'])
+        match_qa(spec_cont_sub, detections, line_list, patt_dict_slit['IDs'], patt_dict_slit['scores'])
 
     # Use only the perfect IDs
     iperfect = np.array(patt_dict_slit['scores']) != 'Perfect'
@@ -995,7 +1202,7 @@ class ArchiveReid:
             # Add the patt_dict and wv_calib to the output dicts
             self.wv_calib[str(slit)] = copy.deepcopy(final_fit)
             if self.debug_fits:
-                qa.arc_fit_qa(self.wv_calib[str(slit)])
+                arc_fit_qa(self.wv_calib[str(slit)])
 
         # Print the final report of all lines
         self.report_final()
@@ -1667,7 +1874,7 @@ class HolyGrail:
 
             if self._debug:
                 tmp_list = table.vstack([self._line_lists, self._unknwns])
-                qa.match_qa(self._spec[:, bs], bsdet, tmp_list,patt_dict['IDs'], patt_dict['scores'])
+                match_qa(self._spec[:, bs], bsdet, tmp_list,patt_dict['IDs'], patt_dict['scores'])
 
             # Use only the perfect IDs
             iperfect = np.array(patt_dict['scores']) != 'Perfect'
@@ -2336,7 +2543,7 @@ class HolyGrail:
 
                 # Plot
                 tmp_list = vstack([self._line_lists, self._unknwns])
-                qa.match_qa(self._spec[:, slit], use_tcent, tmp_list,
+                match_qa(self._spec[:, slit], use_tcent, tmp_list,
                             self._all_patt_dict[str(slit)]['IDs'], self._all_patt_dict[str(slit)]['scores'],
                             outfile=self._outroot + slittxt + '.pdf')
                 msgs.info("Wrote: {:s}".format(self._outroot + slittxt + '.pdf'))
@@ -2481,3 +2688,5 @@ def results_kdtree_nb(use_tcent, wvdata, res, residx, dindex, lindex, nindx, npi
         dind[x, :] = dindex[residx[x], :]
         lind[x, :] = lindex[res[x], :]
     return dind, lind, wvcent, wvdisp
+
+
