+++ pypeit/ginga.py
@@ -1,7 +1,5 @@
 """ Module for ginga routines.  Mainly for debugging
 """
-from __future__ import (print_function, absolute_import, division, unicode_literals)
-
 try:
     basestring
 except NameError:
