+++ pypeit/pypmsgs.py
@@ -390,3 +390,6 @@ class Messages:
         self._white_BL = ''
         self._black_YL = ''
         self._yellow_BK = ''
+
+
+
